split out settings files
