+++ app/settings/__init__.py
@@ -1,0 +1,11 @@
+from .base import *
+from .logging import *
+
+
+# dev settings
+try:
+    from .helpers import dev_mode
+    if dev_mode():
+        from .dev import *
+except ImportError:
+    pass
