+++ app/templatetags/app_filters.py
@@ -1,0 +1,9 @@
+from django.template import Library
+
+
+register = Library()
+
+
+@register.filter(name='times')
+def times(number):
+    return range(number)
