+++ docker.env
@@ -1,3 +1,4 @@
+DEV_MODE=True
 DJANGO_SETTINGS_MODULE=app.settings
 ALLOWED_HOSTS=*
 EMAIL_HOST=mail
